Move GBC::Scale_Rotate implementation to gbuffer.cpp from RA
